+++ tests/Unit/Services/Events/ClientTest.php
@@ -29,6 +29,7 @@ use Shopgate\ConnectSdk\Services\Events\Client;
 use Shopgate\ConnectSdk\Services\Events\Connector\Entities\Base;
 use Shopgate\ConnectSdk\Services\Events\Connector\Entities\Catalog;
 use Shopgate\ConnectSdk\Services\Events\DTO\V1\Payload\Catalog\Category as CategoryDto;
+use Shopgate\ConnectSdk\Services\Events\DTO\V1\Payload\Catalog\Product as ProductDto;
 
 /**
  * @coversDefaultClass \Shopgate\ConnectSdk\Services\Events\Client
@@ -70,19 +71,27 @@ class ClientTest extends TestCase
         $mock             = $this->httpClient->getMock();
         $subjectUnderTest = new Client(['http_client' => $mock]);
         /** @noinspection PhpParamsInspection */
-        $mock->expects($this->exactly(6))->method('request');
+        $mock->expects($this->exactly(14))->method('request');
         $subjectUnderTest->catalog->updateCategory(1, new CategoryDto());
         $subjectUnderTest->catalog->updateCategory(1, new CategoryDto(), [Base::KEY_TYPE => Base::SYNC]);
         $subjectUnderTest->catalog->createCategory(new CategoryDto());
         $subjectUnderTest->catalog->createCategory(new CategoryDto(), [Base::KEY_TYPE => Base::SYNC]);
         $subjectUnderTest->catalog->deleteCategory('1');
         $subjectUnderTest->catalog->deleteCategory('1', [Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->updateProduct(1, new ProductDto());
+        $subjectUnderTest->catalog->updateProduct(1, new ProductDto(), [Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->createProduct(new ProductDto());
+        $subjectUnderTest->catalog->createProduct(new ProductDto(), [Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->deleteProduct('1');
+        $subjectUnderTest->catalog->deleteProduct('1', [Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->getProduct([Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->getProduct([]);
     }
 
     /**
-     * Testing direct calls and service rewrites
+     * Testing direct calls and service rewrites for categories
      */
-    public function testDirectCatalogActions()
+    public function testDirectCatalogCategoryActions()
     {
         $entityId         = 1;
         $defaultMeta      = ['service' => 'catalog', Base::KEY_TYPE => Base::SYNC];
@@ -135,5 +144,75 @@ class ClientTest extends TestCase
         );
         $subjectUnderTest->catalog->deleteCategory($entityId, ['service' => 'test', Base::KEY_TYPE => Base::SYNC]);
         $subjectUnderTest->catalog->createCategory($payload, ['service' => 'test', Base::KEY_TYPE => Base::SYNC]);
+    }
+
+    /**
+     * Testing direct calls and service rewrites for products
+     */
+    public function testDirectCatalogProductActions()
+    {
+        $entityId         = 1;
+        $defaultMeta      = ['service' => 'catalog', Base::KEY_TYPE => Base::SYNC];
+        $mock             = $this->httpClient->getMock();
+        $subjectUnderTest = new Client(['http_client' => $mock]);
+        /** @noinspection PhpParamsInspection */
+        $mock->expects($this->exactly(8))->method('request')->withConsecutive(
+            [
+                $this->equalTo('post'),
+                $this->equalTo('products/' . $entityId),
+                ['query' => $defaultMeta, 'json' => '{}']
+            ],
+            [
+                $this->equalTo('delete'),
+                $this->equalTo('products/' . $entityId),
+                ['query' => $defaultMeta, 'json' => '{}']
+            ],
+            [
+                $this->equalTo('post'),
+                $this->equalTo('products'),
+                ['query' => $defaultMeta, 'json' => '{"products":[[]]}']
+            ],
+            [
+                $this->equalTo('post'),
+                $this->equalTo('products/' . $entityId),
+                ['query' => ['service' => 'test', Base::KEY_TYPE => Base::SYNC], 'json' => '{}']
+            ],
+            [
+                $this->equalTo('delete'),
+                $this->equalTo('products/' . $entityId),
+                ['query' => ['service' => 'test', Base::KEY_TYPE => Base::SYNC], 'json' => '{}']
+            ],
+            [
+                $this->equalTo('post'),
+                $this->equalTo('products'),
+                ['query' => ['service' => 'test', Base::KEY_TYPE => Base::SYNC], 'json' => '{"products":[[]]}']
+            ],
+            [
+                $this->equalTo('get'),
+                $this->equalTo('products'),
+                ['query' => ['service' => 'test', Base::KEY_TYPE => Base::SYNC], 'json' => '{}']
+            ],
+            [
+                $this->equalTo('get'),
+                $this->equalTo('products'),
+                ['query' => ['service' => 'test'], 'json' => '{}']
+            ]
+        );
+
+        $payload = new ProductDto();
+        $subjectUnderTest->catalog->updateProduct($entityId, $payload, [Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->deleteProduct($entityId, [Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->createProduct($payload, [Base::KEY_TYPE => Base::SYNC]);
+
+        // rewriting service via direct call
+        $subjectUnderTest->catalog->updateProduct(
+            $entityId,
+            $payload,
+            ['service' => 'test', Base::KEY_TYPE => Base::SYNC]
+        );
+        $subjectUnderTest->catalog->deleteProduct($entityId, ['service' => 'test', Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->createProduct($payload, ['service' => 'test', Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->getProduct(['service' => 'test', Base::KEY_TYPE => Base::SYNC]);
+        $subjectUnderTest->catalog->getProduct(['service' => 'test']);
     }
 }