+++ CHANGELOG.md
@@ -5,8 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- updateProduct event & direct logic
+- createProduct event & direct logic
+- deleteProduct event & direct logic
+- getProduct direct logic
+
 ### Fixed
 - direct call endpoint
+- most Guzzle HTTP options are now allowed
 
 ## 0.1.0 - 2019-05-10
 ### Added